+++ docs/README.html
@@ -13,7 +13,7 @@
     <script src="_static/jquery.js"></script>
     <script src="_static/underscore.js"></script>
     <script src="_static/doctools.js"></script>
-    <script defer="defer" src="https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js"></script>
+    <script async="async" src="https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js"></script>
     <script>window.MathJax = {"options": {"processHtmlClass": "tex2jax_process|mathjax_process|math|output_area"}}</script>
     <link rel="index" title="Index" href="genindex.html" />
     <link rel="search" title="Search" href="search.html" />
@@ -38,8 +38,8 @@
   <section class="tex2jax_ignore mathjax_ignore" id="a-brief-guide-to-acuitylite">
 <h1>A brief guide to Acuitylite<a class="headerlink" href="#a-brief-guide-to-acuitylite" title="Permalink to this headline">¶</a></h1>
 <p>Acuitylite is an end-to-end neural-network deployment tool for embedded systems.<br/>
-Acuitylite support converting caffe/tflite model to TIM-VX case.
-In addition, Acuitylite support asymmetric uint8 qutization.</p>
+Acuitylite support converting caffe/onnx/tflite models to TIM-VX/TFLite cases.
+In addition, Acuitylite support asymmetric uint8 and symmetric int8 quantization.</p>
 <section id="system-requirement">
 <h2>System Requirement<a class="headerlink" href="#system-requirement" title="Permalink to this headline">¶</a></h2>
 <ul class="simple">
@@ -55,7 +55,7 @@ In addition, Acuitylite support asymmetric uint8 qutization.</p>
 </section>
 <section id="document">
 <h2>Document<a class="headerlink" href="#document" title="Permalink to this headline">¶</a></h2>
-<div class="highlight-none notranslate"><div class="highlight"><pre><span></span>Reference: https://verisilicon.github.io/acuitylite/
+<div class="highlight-none notranslate"><div class="highlight"><pre><span></span>Reference: https://verisilicon.github.io/acuitylite
 </pre></div>
 </div>
 </section>
@@ -64,16 +64,21 @@ In addition, Acuitylite support asymmetric uint8 qutization.</p>
 <ul class="simple">
 <li><p>Importer:
 <a class="reference external" href="https://github.com/BVLC/caffe">Caffe</a>,
+<a class="reference external" href="https://github.com/onnx/onnx">Onnx</a>,
 <a class="reference external" href="https://github.com/tensorflow/tensorflow/tree/master/tensorflow/lite">TFLite</a></p></li>
-<li><p>Exporter:  <a class="reference external" href="https://github.com/VeriSilicon/TIM-VX">TIM-VX</a></p></li>
+<li><p>Exporter:
+<a class="reference external" href="https://github.com/tensorflow/tensorflow/tree/master/tensorflow/lite">TFLite</a>,
+<a class="reference external" href="https://github.com/VeriSilicon/TIM-VX">TIM-VX</a></p></li>
 </ul>
+<p>Tips: You can export a TFLite app and using <a class="reference external" href="https://github.com/VeriSilicon/tflite-vx-delegate">tflite-vx-delegate</a>
+to run on TIM-VX if the exported TIM-VX app does not meet your requirements.</p>
 </section>
 <section id="how-to-run-tim-vx-case">
 <h2>How to run TIM-VX case<a class="headerlink" href="#how-to-run-tim-vx-case" title="Permalink to this headline">¶</a></h2>
 <p>The exported TIM-VX case supports both make and cmake.<br/>
 Please set environment for build and run case:<br/></p>
 <ul class="simple">
-<li><p>TIM_VX_DIR=/path/to/tim-vx/build</p></li>
+<li><p>TIM_VX_DIR=/path/to/tim-vx/build/install</p></li>
 <li><p>VIVANTE_SDK_DIR=/path/to/tim-vx/prebuilt-sdk/x86_64_linux</p></li>
 <li><p>LD_LIBRARY_PATH=<span class="math notranslate nohighlight">\(TIM_VX_DIR/lib:\)</span>VIVANTE_SDK_DIR/lib</p></li>
 </ul>
@@ -122,6 +127,7 @@ You can refer <a class="reference external" href="https://github.com/VeriSilicon
 <p class="caption" role="heading"><span class="caption-text">Demo</span></p>
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="demo_caffe.html">demo.caffe</a></li>
+<li class="toctree-l1"><a class="reference internal" href="demo_onnx.html">demo.onnx</a></li>
 <li class="toctree-l1"><a class="reference internal" href="demo_tflite.html">demo.tflite</a></li>
 </ul>
 
@@ -160,7 +166,7 @@ You can refer <a class="reference external" href="https://github.com/VeriSilicon
       &copy;2021, verisilicon.
       
       |
-      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.3.0</a>
+      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.2.0</a>
       &amp; <a href="https://github.com/bitprophet/alabaster">Alabaster 0.7.12</a>
       
       |
