+++ docs/_sources/demo_onnx.rst.txt
@@ -1,0 +1,5 @@
+demo.onnx
+=========
+
+.. literalinclude:: ../demo/squeezenet_onnx_test.py
+    :language: python
