+++ docs/_sources/demo_tflite.rst.txt
@@ -7,5 +7,5 @@ Download model::
 
 wget https://storage.googleapis.com/download.tensorflow.org/models/mobilenet_v1_2018_08_02/mobilenet_v1_1.0_224_quant.tgz
 
-.. literalinclude:: ../demo/mobilenet_tflite.py
+.. literalinclude:: ../demo/mobilenet_tflite_test.py
     :language: python
