+++ docs/_sources/exporter.tim-vx.rst.txt
@@ -1,0 +1,7 @@
+
+.. autoclass:: exporter.TimVxExporter
+
+    .. automethod:: __init__
+
+    .. automethod:: export
+
