+++ docs/_sources/importer.onnxloader.rst.txt
@@ -1,0 +1,7 @@
+
+.. autoclass:: importer.OnnxLoader
+
+    .. automethod:: __init__
+
+    .. automethod:: load
+
