--- docs/_sources/importer.tflite.rst.txt
@@ -1,6 +1,0 @@
-
-.. autoclass:: importer.TFliteLoader
-
-    .. automethod:: __init__
-
-    .. automethod:: load
