+++ docs/_sources/index.rst.txt
@@ -7,8 +7,8 @@ Welcome to acuitylite's documentation!
 =======================================
 
 Acuitylite is an end-to-end neural-network deployment tool for embedded systems.
-Acuitylite support converting caffe/tflite model to TIM-VX case.
-In addition, Acuitylite support asysmetric uint8 qutization.
+Acuitylite support converting caffe/onnx/tflite model to TIM-VX/TFLite case.
+In addition, Acuitylite support asysmetric uint8 and sysmetric int8 quantization.
 
 Indices and tables
 ==================
@@ -33,6 +33,7 @@ Indices and tables
    :caption: Demo
 
    demo_caffe
+   demo_onnx
    demo_tflite
 
 Support
