+++ docs/_static/basic.css
@@ -731,9 +731,8 @@ dl.glossary dt {
 
 .classifier:before {
     font-style: normal;
-    margin: 0 0.5em;
+    margin: 0.5em;
     content: ":";
-    display: inline-block;
 }
 
 abbr, acronym {
