+++ docs/acuitymodel.html
@@ -33,7 +33,7 @@
             
   <dl class="py class">
 <dt class="sig sig-object py" id="acuitymodel.AcuityModel">
-<em class="property"><span class="pre">class</span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="pre">acuitymodel.</span></span><span class="sig-name descname"><span class="pre">AcuityModel</span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">net</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#acuitymodel.AcuityModel" title="Permalink to this definition">¶</a></dt>
+<em class="property"><span class="pre">class</span> </em><span class="sig-prename descclassname"><span class="pre">acuitymodel.</span></span><span class="sig-name descname"><span class="pre">AcuityModel</span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">net</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#acuitymodel.AcuityModel" title="Permalink to this definition">¶</a></dt>
 <dd><p>Support quantization, inference, exporting TIM_VX application for acuitylite network.</p>
 <dl class="py method">
 <dt class="sig sig-object py" id="acuitymodel.AcuityModel.__init__">
@@ -118,6 +118,7 @@ If set it will improve quantization accuracy, but take more time.</p></li>
 <p class="caption" role="heading"><span class="caption-text">Demo</span></p>
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="demo_caffe.html">demo.caffe</a></li>
+<li class="toctree-l1"><a class="reference internal" href="demo_onnx.html">demo.onnx</a></li>
 <li class="toctree-l1"><a class="reference internal" href="demo_tflite.html">demo.tflite</a></li>
 </ul>
 
@@ -154,7 +155,7 @@ If set it will improve quantization accuracy, but take more time.</p></li>
       &copy;2021, verisilicon.
       
       |
-      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.3.0</a>
+      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.2.0</a>
       &amp; <a href="https://github.com/bitprophet/alabaster">Alabaster 0.7.12</a>
       
       |
