+++ docs/demo_tflite.html
@@ -15,7 +15,7 @@
     <script src="_static/doctools.js"></script>
     <link rel="index" title="Index" href="genindex.html" />
     <link rel="search" title="Search" href="search.html" />
-    <link rel="prev" title="demo.caffe" href="demo_caffe.html" />
+    <link rel="prev" title="demo.onnx" href="demo_onnx.html" />
    
   <link rel="stylesheet" href="_static/custom.css" type="text/css" />
   
@@ -37,10 +37,11 @@
 <p><strong>ATTENTION: No need to quantize using acuity lite for quant model</strong></p>
 <p>Download model:</p>
 <p>wget <a class="reference external" href="https://storage.googleapis.com/download.tensorflow.org/models/mobilenet_v1_2018_08_02/mobilenet_v1_1.0_224_quant.tgz">https://storage.googleapis.com/download.tensorflow.org/models/mobilenet_v1_2018_08_02/mobilenet_v1_1.0_224_quant.tgz</a></p>
-<div class="highlight-python notranslate"><div class="highlight"><pre><span></span><span class="kn">import</span> <span class="nn">tensorflow</span> <span class="k">as</span> <span class="nn">tf</span>
+<div class="highlight-python notranslate"><div class="highlight"><pre><span></span><span class="ch">#!/usr/bin/env python3</span>
+<span class="kn">import</span> <span class="nn">tensorflow</span> <span class="k">as</span> <span class="nn">tf</span>
 <span class="kn">import</span> <span class="nn">numpy</span> <span class="k">as</span> <span class="nn">np</span>
 <span class="kn">from</span> <span class="nn">acuitylib.lite.acuitymodel</span> <span class="kn">import</span> <span class="n">AcuityModel</span>
-<span class="kn">from</span> <span class="nn">acuitylib.lite.importer</span> <span class="kn">import</span> <span class="n">TFliteLoader</span>
+<span class="kn">from</span> <span class="nn">acuitylib.lite.importer</span> <span class="kn">import</span> <span class="n">TFLiteLoader</span>
 <span class="kn">from</span> <span class="nn">acuitylib.lite.exporter</span> <span class="kn">import</span> <span class="n">TimVxExporter</span>
 
 <span class="c1"># wget https://storage.googleapis.com/download.tensorflow.org/</span>
@@ -58,10 +59,10 @@
         <span class="n">arr</span> <span class="o">=</span> <span class="p">(</span><span class="n">arr</span><span class="o">-</span><span class="mf">128.0</span><span class="p">)</span><span class="o">/</span><span class="mf">128.0</span>  <span class="c1"># preprocess</span>
         <span class="k">yield</span> <span class="n">np</span><span class="o">.</span><span class="n">array</span><span class="p">(</span><span class="n">arr</span><span class="p">,</span> <span class="n">dtype</span><span class="o">=</span><span class="n">np</span><span class="o">.</span><span class="n">float32</span><span class="p">)</span>
 
-<span class="k">def</span> <span class="nf">main</span><span class="p">():</span>
+<span class="k">def</span> <span class="nf">test_tflite_mobilenet</span><span class="p">():</span>
     <span class="c1"># no need to quantize using acuity lite for quant model</span>
     <span class="c1"># load tflite quant model</span>
-    <span class="n">importtflite</span> <span class="o">=</span> <span class="n">TFliteLoader</span><span class="p">(</span><span class="n">mobilenet</span><span class="p">)</span>
+    <span class="n">importtflite</span> <span class="o">=</span> <span class="n">TFLiteLoader</span><span class="p">(</span><span class="n">mobilenet</span><span class="p">)</span>
     <span class="n">acuitynet</span> <span class="o">=</span> <span class="n">importtflite</span><span class="o">.</span><span class="n">load</span><span class="p">()</span>
     <span class="n">quantmodel</span> <span class="o">=</span> <span class="n">AcuityModel</span><span class="p">(</span><span class="n">acuitynet</span><span class="p">)</span>
 
@@ -76,7 +77,7 @@
     <span class="n">TimVxExporter</span><span class="p">(</span><span class="n">quantmodel</span><span class="p">)</span><span class="o">.</span><span class="n">export</span><span class="p">(</span><span class="s1">&#39;export_timvx/quant/mobilenet&#39;</span><span class="p">)</span>
 
 <span class="k">if</span> <span class="vm">__name__</span> <span class="o">==</span> <span class="s1">&#39;__main__&#39;</span><span class="p">:</span>
-    <span class="n">main</span><span class="p">()</span>
+    <span class="n">test_tflite_mobilenet</span><span class="p">()</span>
 </pre></div>
 </div>
 </section>
@@ -109,6 +110,7 @@
 <p class="caption" role="heading"><span class="caption-text">Demo</span></p>
 <ul class="current">
 <li class="toctree-l1"><a class="reference internal" href="demo_caffe.html">demo.caffe</a></li>
+<li class="toctree-l1"><a class="reference internal" href="demo_onnx.html">demo.onnx</a></li>
 <li class="toctree-l1 current"><a class="current reference internal" href="#">demo.tflite</a></li>
 </ul>
 
@@ -116,7 +118,7 @@
 <h3>Related Topics</h3>
 <ul>
   <li><a href="index.html">Documentation overview</a><ul>
-      <li>Previous: <a href="demo_caffe.html" title="previous chapter">demo.caffe</a></li>
+      <li>Previous: <a href="demo_onnx.html" title="previous chapter">demo.onnx</a></li>
   </ul></li>
 </ul>
 </div>
@@ -146,7 +148,7 @@
       &copy;2021, verisilicon.
       
       |
-      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.3.0</a>
+      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.2.0</a>
       &amp; <a href="https://github.com/bitprophet/alabaster">Alabaster 0.7.12</a>
       
       |
