+++ docs/genindex.html
@@ -40,6 +40,7 @@
  | <a href="#E"><strong>E</strong></a>
  | <a href="#I"><strong>I</strong></a>
  | <a href="#L"><strong>L</strong></a>
+ | <a href="#O"><strong>O</strong></a>
  | <a href="#Q"><strong>Q</strong></a>
  | <a href="#T"><strong>T</strong></a>
  
@@ -50,11 +51,15 @@
       <li><a href="acuitymodel.html#acuitymodel.AcuityModel.__init__">__init__() (acuitymodel.AcuityModel method)</a>
 
       <ul>
-        <li><a href="exporter.html#exporter.TimVxExporter.__init__">(exporter.TimVxExporter method)</a>
+        <li><a href="exporter.tflite.html#exporter.TFLiteExporter.__init__">(exporter.TFLiteExporter method)</a>
+</li>
+        <li><a href="exporter.tim-vx.html#exporter.TimVxExporter.__init__">(exporter.TimVxExporter method)</a>
 </li>
         <li><a href="importer.caffeloader.html#importer.CaffeLoader.__init__">(importer.CaffeLoader method)</a>
 </li>
-        <li><a href="importer.tflite.html#importer.TFliteLoader.__init__">(importer.TFliteLoader method)</a>
+        <li><a href="importer.onnxloader.html#importer.OnnxLoader.__init__">(importer.OnnxLoader method)</a>
+</li>
+        <li><a href="importer.tfliteloader.html#importer.TFLiteLoader.__init__">(importer.TFLiteLoader method)</a>
 </li>
       </ul></li>
   </ul></td>
@@ -79,8 +84,12 @@
 <h2 id="E">E</h2>
 <table style="width: 100%" class="indextable genindextable"><tr>
   <td style="width: 33%; vertical-align: top;"><ul>
-      <li><a href="exporter.html#exporter.TimVxExporter.export">export() (exporter.TimVxExporter method)</a>
+      <li><a href="exporter.tflite.html#exporter.TFLiteExporter.export">export() (exporter.TFLiteExporter method)</a>
+
+      <ul>
+        <li><a href="exporter.tim-vx.html#exporter.TimVxExporter.export">(exporter.TimVxExporter method)</a>
 </li>
+      </ul></li>
   </ul></td>
 </tr></table>
 
@@ -98,9 +107,19 @@
       <li><a href="importer.caffeloader.html#importer.CaffeLoader.load">load() (importer.CaffeLoader method)</a>
 
       <ul>
-        <li><a href="importer.tflite.html#importer.TFliteLoader.load">(importer.TFliteLoader method)</a>
+        <li><a href="importer.onnxloader.html#importer.OnnxLoader.load">(importer.OnnxLoader method)</a>
+</li>
+        <li><a href="importer.tfliteloader.html#importer.TFLiteLoader.load">(importer.TFLiteLoader method)</a>
 </li>
       </ul></li>
+  </ul></td>
+</tr></table>
+
+<h2 id="O">O</h2>
+<table style="width: 100%" class="indextable genindextable"><tr>
+  <td style="width: 33%; vertical-align: top;"><ul>
+      <li><a href="importer.onnxloader.html#importer.OnnxLoader">OnnxLoader (class in importer)</a>
+</li>
   </ul></td>
 </tr></table>
 
@@ -119,11 +138,13 @@
 <h2 id="T">T</h2>
 <table style="width: 100%" class="indextable genindextable"><tr>
   <td style="width: 33%; vertical-align: top;"><ul>
-      <li><a href="importer.tflite.html#importer.TFliteLoader">TFliteLoader (class in importer)</a>
+      <li><a href="exporter.tflite.html#exporter.TFLiteExporter">TFLiteExporter (class in exporter)</a>
 </li>
   </ul></td>
   <td style="width: 33%; vertical-align: top;"><ul>
-      <li><a href="exporter.html#exporter.TimVxExporter">TimVxExporter (class in exporter)</a>
+      <li><a href="importer.tfliteloader.html#importer.TFLiteLoader">TFLiteLoader (class in importer)</a>
+</li>
+      <li><a href="exporter.tim-vx.html#exporter.TimVxExporter">TimVxExporter (class in exporter)</a>
 </li>
   </ul></td>
 </tr></table>
@@ -157,6 +178,7 @@
 <p class="caption" role="heading"><span class="caption-text">Demo</span></p>
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="demo_caffe.html">demo.caffe</a></li>
+<li class="toctree-l1"><a class="reference internal" href="demo_onnx.html">demo.onnx</a></li>
 <li class="toctree-l1"><a class="reference internal" href="demo_tflite.html">demo.tflite</a></li>
 </ul>
 
@@ -193,7 +215,7 @@
       &copy;2021, verisilicon.
       
       |
-      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.3.0</a>
+      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.2.0</a>
       &amp; <a href="https://github.com/bitprophet/alabaster">Alabaster 0.7.12</a>
       
     </div>
