+++ docs/importer.caffeloader.html
@@ -33,8 +33,8 @@
             
   <dl class="py class">
 <dt class="sig sig-object py" id="importer.CaffeLoader">
-<em class="property"><span class="pre">class</span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="pre">importer.</span></span><span class="sig-name descname"><span class="pre">CaffeLoader</span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">model</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">weights</span></span><span class="o"><span class="pre">=</span></span><span class="default_value"><span class="pre">None</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#importer.CaffeLoader" title="Permalink to this definition">¶</a></dt>
-<dd><p>Translates a Caffe model to ACUITY formats.</p>
+<em class="property"><span class="pre">class</span> </em><span class="sig-prename descclassname"><span class="pre">importer.</span></span><span class="sig-name descname"><span class="pre">CaffeLoader</span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">model</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">weights</span></span><span class="o"><span class="pre">=</span></span><span class="default_value"><span class="pre">None</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#importer.CaffeLoader" title="Permalink to this definition">¶</a></dt>
+<dd><p>Translates a Caffe model to Acuity formats.</p>
 <dl class="py method">
 <dt class="sig sig-object py" id="importer.CaffeLoader.__init__">
 <span class="sig-name descname"><span class="pre">__init__</span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">model</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">weights</span></span><span class="o"><span class="pre">=</span></span><span class="default_value"><span class="pre">None</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#importer.CaffeLoader.__init__" title="Permalink to this definition">¶</a></dt>
@@ -54,7 +54,7 @@
 <dd><p>Load Caffe model.</p>
 <dl class="field-list simple">
 <dt class="field-odd">Returns</dt>
-<dd class="field-odd"><p>Acuitylite network.</p>
+<dd class="field-odd"><p>Acuity network.</p>
 </dd>
 </dl>
 </dd></dl>
@@ -90,6 +90,7 @@
 <p class="caption" role="heading"><span class="caption-text">Demo</span></p>
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="demo_caffe.html">demo.caffe</a></li>
+<li class="toctree-l1"><a class="reference internal" href="demo_onnx.html">demo.onnx</a></li>
 <li class="toctree-l1"><a class="reference internal" href="demo_tflite.html">demo.tflite</a></li>
 </ul>
 
@@ -126,7 +127,7 @@
       &copy;2021, verisilicon.
       
       |
-      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.3.0</a>
+      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.2.0</a>
       &amp; <a href="https://github.com/bitprophet/alabaster">Alabaster 0.7.12</a>
       
       |
