+++ docs/index.html
@@ -35,8 +35,8 @@
   <section id="welcome-to-acuitylite-s-documentation">
 <h1>Welcome to acuitylite’s documentation!<a class="headerlink" href="#welcome-to-acuitylite-s-documentation" title="Permalink to this headline">¶</a></h1>
 <p>Acuitylite is an end-to-end neural-network deployment tool for embedded systems.
-Acuitylite support converting caffe/tflite model to TIM-VX case.
-In addition, Acuitylite support asysmetric uint8 qutization.</p>
+Acuitylite support converting caffe/onnx/tflite model to TIM-VX/TFLite case.
+In addition, Acuitylite support asysmetric uint8 and sysmetric int8 quantization.</p>
 </section>
 <section id="indices-and-tables">
 <h1>Indices and tables<a class="headerlink" href="#indices-and-tables" title="Permalink to this headline">¶</a></h1>
@@ -74,6 +74,7 @@ In addition, Acuitylite support asysmetric uint8 qutization.</p>
 <p class="caption" role="heading"><span class="caption-text">Demo</span></p>
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="demo_caffe.html">demo.caffe</a></li>
+<li class="toctree-l1"><a class="reference internal" href="demo_onnx.html">demo.onnx</a></li>
 <li class="toctree-l1"><a class="reference internal" href="demo_tflite.html">demo.tflite</a></li>
 </ul>
 </div>
@@ -111,6 +112,7 @@ In addition, Acuitylite support asysmetric uint8 qutization.</p>
 <p class="caption" role="heading"><span class="caption-text">Demo</span></p>
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="demo_caffe.html">demo.caffe</a></li>
+<li class="toctree-l1"><a class="reference internal" href="demo_onnx.html">demo.onnx</a></li>
 <li class="toctree-l1"><a class="reference internal" href="demo_tflite.html">demo.tflite</a></li>
 </ul>
 
@@ -148,7 +150,7 @@ In addition, Acuitylite support asysmetric uint8 qutization.</p>
       &copy;2021, verisilicon.
       
       |
-      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.3.0</a>
+      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.2.0</a>
       &amp; <a href="https://github.com/bitprophet/alabaster">Alabaster 0.7.12</a>
       
       |
