+++ docs/quantizer.html
@@ -33,11 +33,16 @@
             
   <dl class="py class">
 <dt class="sig sig-object py" id="quantizer.QuantizeType">
-<em class="property"><span class="pre">class</span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="pre">quantizer.</span></span><span class="sig-name descname"><span class="pre">QuantizeType</span></span><a class="headerlink" href="#quantizer.QuantizeType" title="Permalink to this definition">¶</a></dt>
-<dd><p>Quantizer type for AcuityModel quantize method quantizer param</p>
+<em class="property"><span class="pre">class</span> </em><span class="sig-prename descclassname"><span class="pre">quantizer.</span></span><span class="sig-name descname"><span class="pre">QuantizeType</span></span><a class="headerlink" href="#quantizer.QuantizeType" title="Permalink to this definition">¶</a></dt>
+<dd><p>Quantizer type for AcuityModel quantize method quantizer param,
+asymu8 = ASYMMETRIC_AFFINE_UINT8,
+pcq_symi8 = PERCHANNEL_SYMMETRIC_AFFINE_INT8</p>
 <dl class="field-list simple">
 <dt class="field-odd">Members</dt>
 <dd class="field-odd"><p>asymu8</p>
+</dd>
+<dt class="field-even">Members</dt>
+<dd class="field-even"><p>pcq_symi8</p>
 </dd>
 </dl>
 </dd></dl>
@@ -71,6 +76,7 @@
 <p class="caption" role="heading"><span class="caption-text">Demo</span></p>
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="demo_caffe.html">demo.caffe</a></li>
+<li class="toctree-l1"><a class="reference internal" href="demo_onnx.html">demo.onnx</a></li>
 <li class="toctree-l1"><a class="reference internal" href="demo_tflite.html">demo.tflite</a></li>
 </ul>
 
@@ -107,7 +113,7 @@
       &copy;2021, verisilicon.
       
       |
-      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.3.0</a>
+      Powered by <a href="http://sphinx-doc.org/">Sphinx 4.2.0</a>
       &amp; <a href="https://github.com/bitprophet/alabaster">Alabaster 0.7.12</a>
       
       |
