+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["README","acuitymodel","api","demo_caffe","demo_tflite","exporter","importer.caffeloader","importer.tflite","index","quantizer"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":4,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":3,"sphinx.domains.rst":2,"sphinx.domains.std":2,sphinx:56},filenames:["README.md","acuitymodel.rst","api.rst","demo_caffe.rst","demo_tflite.rst","exporter.rst","importer.caffeloader.rst","importer.tflite.rst","index.rst","quantizer.rst"],objects:{"acuitymodel.AcuityModel":[[1,1,1,"","__init__"],[1,1,1,"","inference"],[1,1,1,"","quantize"]],"exporter.TimVxExporter":[[5,1,1,"","__init__"],[5,1,1,"","export"]],"importer.CaffeLoader":[[6,1,1,"","__init__"],[6,1,1,"","load"]],"importer.TFliteLoader":[[7,1,1,"","__init__"],[7,1,1,"","load"]],acuitymodel:[[1,0,1,"","AcuityModel"]],exporter:[[5,0,1,"","TimVxExporter"]],importer:[[6,0,1,"","CaffeLoader"],[7,0,1,"","TFliteLoader"]],quantizer:[[9,0,1,"","QuantizeType"]]},objnames:{"0":["py","class","Python class"],"1":["py","method","Python method"]},objtypes:{"0":"py:class","1":"py:method"},terms:{"0":[3,4],"04":0,"0_224_quant":4,"1":[1,3,4,7],"1000":1,"12":7,"128":4,"2":[3,4],"20":0,"224":[4,7],"255":3,"28":3,"3":[4,7],"500":1,"64":0,"697":4,"8":0,"813":4,"9":4,"99":3,"case":[3,4,8],"class":[1,5,6,7,9],"do":[1,7],"export":[0,1,3,4,5,8],"import":[0,3,4,6,7,8],"int":1,"return":[1,5,6,7],A:8,For:[5,7],If:1,In:[0,8],No:4,Such:7,The:[0,1,5,6,7],_:[3,4],__init__:[1,5,6,7],__main__:[3,4],__name__:[3,4],accuraci:1,acuiti:[1,4,6],acuitylib:[3,4],acuitylit:[1,5,6,7],acuitymodel:[1,3,4,5,8,9],acuitynet:[3,4],addit:[0,8],again:7,all:7,an:[0,1,5,7,8],applic:[1,5],ar:1,argument:7,arr:[3,4],arrai:[3,4],assert:[3,4],astyp:[3,4],asymmetr:0,asymu8:[1,3,9],asysmetr:8,attent:[0,4],batch:1,batch_siz:[1,3,4],been:1,bit:0,bool:1,both:0,brief:8,build:0,caff:[0,6,8],caffe_model:3,caffeload:[3,6],caffemodel:[3,6],can:0,cmake:0,com:[0,4,5,8],convert:[0,1,8],creat:[0,8],data:[1,3,4],data_list:[3,4],decode_imag:[3,4],def:[3,4],deploy:[0,8],directori:5,download:4,dtype:[3,4],email:[0,8],embed:[0,8],end:[0,8],enumer:[3,4],environ:0,error:1,exampl:[5,7],export_timvx:[3,4,5],extens:7,fals:1,file:[0,5,6,7],flatbuff:7,float16:3,float32:[3,4],format:[6,7],framework:8,from:[1,3,4],gener:[1,3,4,5],get_data:[3,4],github:[0,5,8],googleapi:4,graph:7,have:1,header:[0,7],how:8,http:[0,4,5],i:[3,4],identifi:7,imag:[1,3,4],image0:[3,4],image1:[3,4],importcaff:3,importtflit:4,improv:1,includ:[0,5],index:8,infer:[1,3,4],input:[1,7],input_size_list:7,instal:8,io:[0,3,4],issu:[0,8],item:[3,4],iter:[1,3,4],jpg:[3,4],label:4,ld_library_path:0,lenet:[3,5],lib:0,linux:0,list:[1,7],lite:[1,3,4,7],load:[3,4,6,7],lt:0,main:[3,4],make:0,member:9,method:9,minim:1,minimize_layer_error:1,ml_support:[0,8],mobilenet:4,mobilenet_v1_1:4,mobilenet_v1_2018_08_02:4,model:[0,1,3,4,5,6,7,8],more:1,need:[0,1,4,7],net:1,network:[0,1,6,7,8],neural:[0,8],none:[6,7],now:1,np:[3,4],number:1,numpi:[3,4],object:[1,5],omit:7,optim:7,org:4,os:0,out:[3,4],output:[5,7],output_point_1:7,output_point_2:7,page:8,param:[7,9],paramet:[1,5,6,7],path:[0,6,7],per:1,perform:1,pictur:1,pip:0,pleas:0,png:3,point:7,prebuilt:0,prefix:5,preprocess:[3,4],prototxt:[3,6],python3:0,python:0,quant:4,quantiz:[1,3,4,7,8,9],quantizetyp:[1,3,9],quantmodel:4,qutiz:[0,8],read_fil:[3,4],recommend:0,refer:[0,5],repres:7,requir:[1,7,8],requri:7,reshap:[3,4],result:[1,3,4],run:8,same:7,sampl:1,sdk:0,search:8,set:[0,1,7],shape:7,should:0,size:7,specifi:[5,7],storag:4,subdirectori:5,support:1,syntax:5,system:8,tail:7,take:1,tensor:[1,7],tensorflow:[1,3,4,7],tf:[3,4],tflite:[0,7,8],tfliteload:[4,7],tgz:4,them:7,thi:7,tim:[3,4,5,8],tim_vx:1,tim_vx_dir:0,time:[1,7],timvxexport:[3,4,5],tool:[0,8],translat:[6,7],two:7,type:[1,7,9],ubuntu:0,uint8:[0,3,5,8],us:[4,5,7],usual:1,verisilicon:[0,5,8],version:0,vivante_sdk_dir:0,vx:[3,4,5,8],weight:[3,6],wget:4,when:7,where:5,which:7,x86_64_linux:0,yield:[3,4],you:0},titles:["A brief guide to Acuitylite","&lt;no title&gt;","API","demo.caffe","demo.tflite","&lt;no title&gt;","&lt;no title&gt;","&lt;no title&gt;","Welcome to acuitylite\u2019s documentation!","&lt;no title&gt;"],titleterms:{"case":0,"export":2,"import":2,A:0,acuitylit:[0,8],acuitymodel:2,api:[2,8],brief:0,caff:3,demo:[3,4,8],document:[0,8],framework:0,guid:[0,8],how:0,indic:8,instal:0,quantiz:2,refer:8,requir:0,run:0,s:8,support:[0,8],system:0,tabl:8,tflite:4,tim:0,vx:0,welcom:8}})
+Search.setIndex({docnames:["README","acuitymodel","api","demo_caffe","demo_onnx","demo_tflite","exporter.tflite","exporter.tim-vx","importer.caffeloader","importer.onnxloader","importer.tfliteloader","index","quantizer"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":4,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":3,"sphinx.domains.rst":2,"sphinx.domains.std":2,sphinx:56},filenames:["README.md","acuitymodel.rst","api.rst","demo_caffe.rst","demo_onnx.rst","demo_tflite.rst","exporter.tflite.rst","exporter.tim-vx.rst","importer.caffeloader.rst","importer.onnxloader.rst","importer.tfliteloader.rst","index.rst","quantizer.rst"],objects:{"acuitymodel.AcuityModel":{__init__:[1,1,1,""],inference:[1,1,1,""],quantize:[1,1,1,""]},"exporter.TFLiteExporter":{"export":[6,1,1,""],__init__:[6,1,1,""]},"exporter.TimVxExporter":{"export":[7,1,1,""],__init__:[7,1,1,""]},"importer.CaffeLoader":{__init__:[8,1,1,""],load:[8,1,1,""]},"importer.OnnxLoader":{__init__:[9,1,1,""],load:[9,1,1,""]},"importer.TFLiteLoader":{__init__:[10,1,1,""],load:[10,1,1,""]},acuitymodel:{AcuityModel:[1,0,1,""]},exporter:{TFLiteExporter:[6,0,1,""],TimVxExporter:[7,0,1,""]},importer:{CaffeLoader:[8,0,1,""],OnnxLoader:[9,0,1,""],TFLiteLoader:[10,0,1,""]},quantizer:{QuantizeType:[12,0,1,""]}},objnames:{"0":["py","class","Python class"],"1":["py","method","Python method"]},objtypes:{"0":"py:class","1":"py:method"},terms:{"0":[3,4,5],"04":0,"0_224_quant":5,"1":[1,3,4,5,9,10],"1000":1,"12":[9,10],"128":5,"2":[3,4,5],"20":0,"224":[4,5,9,10],"255":3,"28":3,"3":[4,5,9,10],"500":1,"64":0,"697":5,"8":0,"812":4,"813":[4,5],"9":5,"99":3,"case":[3,4,5,11],"class":[1,6,7,8,9,10,12],"do":[1,10],"export":[0,1,3,4,5,6,7,11],"import":[0,3,4,5,8,9,10,11],"int":1,"return":[1,6,7,8,9,10],A:11,For:[6,7,9,10],If:1,In:[0,11],No:5,Such:[9,10],The:[0,1,6,7,8,9,10],_:[3,4,5],__init__:[1,6,7,8,9,10],__main__:[3,4,5],__name__:[3,4,5],accuraci:1,acuiti:[1,5,6,7,8,9,10],acuitylib:[3,4,5],acuitylit:[1,9,10],acuitymodel:[1,3,4,5,6,7,11,12],acuitynet:[3,5],addit:[0,11],again:10,all:[9,10],an:[0,1,6,7,10,11],app:0,applic:[1,6,7],ar:1,argmax:4,argument:[9,10],arr:[3,4,5],arrai:[3,4,5],assert:[3,4,5],astyp:[3,5],asymmetr:0,asymmetric_affine_uint8:12,asymu8:[1,4,12],asysmetr:11,attent:[0,5],batch:1,batch_siz:[1,3,4,5],been:1,bin:[3,4,5],bit:0,bool:1,both:0,brief:11,build:0,caff:[0,8,11],caffe_model:3,caffeload:[3,8],caffemodel:[3,8],can:0,cmake:0,com:[0,5,6,7,11],convert:[0,1,11],creat:[0,11],data:[1,3,4,5],data_0:4,data_list:[3,4,5],dataset:4,decode_imag:[3,4,5],def:[3,4,5],deleg:0,deploy:[0,11],directori:[6,7],doe:0,download:5,dtype:[3,4,5],email:[0,11],embed:[0,11],end:[0,11],enumer:[3,4,5],env:[3,4,5],environ:0,error:1,exampl:[6,7,9,10],export_tflit:[3,4,6],export_timvx:[3,4,5,7],extens:10,fals:1,file:[0,6,7,8,9,10],flatbuff:10,flatten:4,float16:[3,4],float32:[3,4,5],format:[8,9,10],framework:11,from:[1,3,4,5],gener:[1,3,4,5,6,7],get_data:[3,4,5],github:[0,6,7,11],googleapi:5,graph:[9,10],have:1,header:[0,9,10],how:11,http:[0,5,6,7],i:[3,4,5],identifi:10,imag:[1,3,4,5],image0:[3,4,5],image1:[3,5],importcaff:3,importtflit:5,improv:1,includ:[0,6,7],index:11,infer:[1,3,4,5],input:[1,4,9,10],input_point_1:[9,10],input_point_2:[9,10],input_size_list:[4,9,10],instal:11,int8:[0,11],io:[0,3,4,5],issu:[0,11],item:[3,4,5],iter:[1,3,4,5],jpg:[3,4,5],label:5,ld_library_path:0,lenet:[3,4,6,7],lib:0,linux:0,list:[1,9,10],lite:[1,3,4,5,6,10],load:[3,4,5,8,9,10],lt:0,make:0,master:6,meet:0,member:12,method:12,minim:1,minimize_layer_error:1,ml_support:[0,11],mobilenet:5,mobilenet_v1_1:5,mobilenet_v1_2018_08_02:5,model:[0,1,3,4,5,6,7,8,9,10,11],more:1,need:[0,1,5,10],net:[1,4],network:[0,1,8,9,10,11],neural:[0,11],none:[8,9,10],now:1,np:[3,4,5],number:1,numpi:[3,4,5],object:[1,6,7],omit:[9,10],onnx:[0,9,11],onnx_model:4,onnxload:[4,9],optim:10,org:5,os:[0,4],out:[3,4,5],output:[4,6,7,9,10],output_point_1:[9,10],output_point_2:[9,10],page:11,param:[10,12],paramet:[1,6,7,8,9,10],path:[0,8,9,10],pcq_symi8:[3,12],per:1,perchannel_symmetric_affine_int8:12,perchanneli8:3,perform:1,pictur:1,pip:0,pleas:0,png:3,point:[9,10],prebuilt:0,prefix:[6,7],prepar:4,preprocess:[3,5],prototxt:[3,8],python3:[0,3,4,5],python:0,quant:5,quantiz:[0,1,3,4,5,10,11,12],quantizetyp:[1,3,4,12],quantmodel:5,read_fil:[3,4,5],recommend:0,refer:[0,6,7],repres:[9,10],requir:[1,9,10,11],requri:[9,10],reshap:[3,4,5],result:[1,3,4,5],run:11,same:10,sampl:1,sdk:0,search:11,set:[0,1,10],shape:[9,10],should:0,size:[9,10],softmaxout_1:4,specifi:[6,7,9,10],squeezenet:4,storag:5,subdirectori:[6,7],support:1,symmetr:0,syntax:[6,7],sysmetr:11,system:11,tail:[9,10],take:1,tensor:[1,9,10],tensorflow:[1,3,4,5,6,10],test_caffe_lenet:3,test_onnx_squeezenet:4,test_tflite_mobilenet:5,tf:[3,4,5],tflite:[0,3,4,6,9,10,11],tfliteexport:[3,4,6],tfliteload:[5,10],tgz:5,them:10,thi:[9,10],tim:[3,4,5,7,11],tim_vx:1,tim_vx_dir:0,time:[1,10],timvxexport:[3,4,5,7],tip:0,tool:[0,11],translat:[8,9,10],transpos:4,tree:6,two:[9,10],type:[1,9,10,12],ubuntu:0,uint8:[0,6,7,11],us:[0,5,6,7,9,10],usr:[3,4,5],usual:1,verisilicon:[0,7,11],version:0,vivante_sdk_dir:0,vx:[3,4,5,7,11],weight:[3,8],wget:5,when:[9,10],where:[6,7],which:[9,10],x86_64_linux:0,yield:[3,4,5],you:0,your:0},titles:["A brief guide to Acuitylite","&lt;no title&gt;","API","demo.caffe","demo.onnx","demo.tflite","&lt;no title&gt;","&lt;no title&gt;","&lt;no title&gt;","&lt;no title&gt;","&lt;no title&gt;","Welcome to acuitylite\u2019s documentation!","&lt;no title&gt;"],titleterms:{"case":0,"export":2,"import":2,A:0,acuitylit:[0,11],acuitymodel:2,api:[2,11],brief:0,caff:3,demo:[3,4,5,11],document:[0,11],framework:0,guid:[0,11],how:0,indic:11,instal:0,onnx:4,quantiz:2,refer:11,requir:0,run:0,s:11,support:[0,11],system:0,tabl:11,tflite:5,tim:0,vx:0,welcom:11}})
